requirements.txt
# requirements.txt
# pip install -r requirements.txt
numpy
scipy
matplotlib
streamlit
pytest
pyyaml
jsonschema